requirements.txt
Flask==2.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
PyDrive==1.3.1
mysql-connector-python==8.0.28
